+++ src/core/Slack/repository/SlackRepositoryInterface.ts
@@ -1,0 +1,5 @@
+import { accountstatusType } from "../model/slack.model";
+
+export interface SlackRepositoryInterface{
+    getAccountStatuses(): Promise<accountstatusType[]>
+}
