+++ src/handlers/removeResignedPeople.ts
@@ -1,6 +1,7 @@
 import { SSM } from '@aws-sdk/client-ssm'
 import { EffortService } from '@src/core/Effort/service/EffortService'
 import { SkillMatrixService } from '@src/core/SkillMatrix/service/SkillMatrixService'
+import { removeResignedPeople } from '@src/core/User/runner/removeResignedPeopleRunner'
 import { ResignedPeopleService } from '@src/core/User/service/ResignedPeopleService'
 import { UserProfileService } from '@src/core/User/service/UserProfileService'
 import { EffortRepository } from '@src/infrastructure/Effort/repository/EffortRepository'
@@ -19,38 +20,22 @@ const resignedPeopleService = new ResignedPeopleService(
 
 export async function handler() {
   const slackClient = new SlackClient(await getSlackToken())
-  const slackUsersStatus = await slackClient.getAccountStatuses()
-  const users = await userService.getAllUserProfiles()
-  users.forEach(async (user) => {
-    const userEmails = [
-      user.uid.toLowerCase(),
-      user.uid
-        .replace('claranet.com', `${user.company}.clara.net`)
-        .toLowerCase(),
-    ]
-    const slackUser = slackUsersStatus.find((u) =>
-      userEmails.includes(u.email.toLowerCase()),
-    )
-    if (slackUser) {
-      if (!slackUser.active) {
-        console.log(`removing ${user.uid}`)
-        await resignedPeopleService.removeResigned(user.uid)
-      }
-    } else {
-      console.warn(`User ${user.uid} not found in slack`)
-    }
-  })
+  await removeResignedPeople(
+    userService.getAllUserProfiles.bind(userService),
+    slackClient.getAccountStatuses.bind(slackClient),
+    resignedPeopleService.removeResigned.bind(resignedPeopleService),
+  )
   return { message: 'done' }
 }
 
 async function getSlackToken(): Promise<string> {
-    const ssm = new SSM()
-    const key = await ssm.getParameter({
-        Name: process.env.SLACK_TOKEN_ARN,
-        WithDecryption: true,
-      })
-      if(!key.Parameter){
-          throw new Error('Slack token not found')
-      }
-      return key.Parameter.Value!
+  const ssm = new SSM()
+  const key = await ssm.getParameter({
+    Name: process.env.SLACK_TOKEN_ARN,
+    WithDecryption: true,
+  })
+  if (!key.Parameter) {
+    throw new Error('Slack token not found')
+  }
+  return key.Parameter.Value!
 }
